1.0.26 - Documentando classes, atributos e métodos das classes que estão estabilizadas. - TimerSystem atualizado para funcionar conforme a Orientação a Objetos. 	- Agora timers lista os temporizadores por ID e tickers ordena por tempo de expiração. 	- Para isso foi implementado uma classe para ordenar os temporizadores: Tickers. 	- Agora o tempo no servidor será trabalhado com int e não mais long. - Corrigindo o nome de FileDecriptor que é FileDescriptor, listener corrigido também.

+++ JRagnarok Emulador/src/org/diverproject/jragnarok/server/TimerListener.java
@@ -1,6 +1,22 @@
 package org.diverproject.jragnarok.server;
 
+/**
+ * Listener para Temporizador
+ *
+ * Listener usado para executar comandos/operações de um temporizador.
+ * Eles são vinculados ao temporizadores para definir a ação do mesmo.
+ *
+ * @author Andrew
+ */
+
 public interface TimerListener
 {
-	void onCall(Timer timer, long tick);
+	/**
+	 * Procedimento chamado internamente pelo Sistema de Temporizadores.
+	 * A chamada ocorre somente quando o tempo do temporizador timer expirado.
+	 * @param timer referência do temporizador que detêm esse listener.
+	 * @param tick momento em que essa chama está sendo feita no sistema.
+	 */
+
+	void onCall(Timer timer, int tick);
 }
